+++ test/resources/language/elixir/lib/infos.ex
@@ -1,0 +1,5 @@
+defmodule Infos do
+  def give_infos() do
+    IO.puts "Hello World"
+  end
+end